backend/requirements.txt
flask
pymongo
redis
flask-socketio
flask-cors
werkzeug
pyjwt
psutil
requests
python-dotenv
polygon-api-client
finnhub-python
google-generativeai
autogen
schedule
numpy
pandas
